Add Scanner option to expand variations (part 1 of 2)
This commit is part of a series to add an optional ability to expand
variations when utilizing Scanner to iterate over a .pgn file.

This commit specifically introduces ScannerOpts to enable consumers of
Scanner to specify different scanning behavior. The first intended use
is ScannerOpts.ExpandVariations. Currently Scanner will strip out any
variations that are present in a game defined within a .pgn file. This
new option (non-default) when specified will instead allow Scanner
consumers to specify that when iterating over a .pgn it should create
a new instance of a chess game for every variation
encountered. Additionally this commit refactors Scanner internals to
deal with the fact that decoding a single game may return a plural
number of Game instances rather than a singular. Code to actually
actually make this plural will be in a followup commit.