Add dark mode & perspective to Board.Draw()
This commit adds Board.Draw2(perspective Color, darkMode bool). This
is similar to Board.Draw() except that it allows the caller to specify
from which perspective to display and whether the display should be
reversed colors. (In a terminal window configured with white foreground
text and a black background, the current Board.Draw() displays the
opposite color).